refactor: FTA audit batch 7 — decompose usePhotoActions, ExercisePickerSheet, BodyCards, recommend (BLD-332)
Extract sub-modules to reduce file complexity:
- lib/photo-processing.ts + hooks/usePhotoPermissions.ts from usePhotoActions (375→309L)
- components/exercise-picker/ExerciseItem.tsx from ExercisePickerSheet (352→291L)
- components/progress/ChartCard.tsx + GoalsCard.tsx from BodyCards (333→170L)
- components/onboarding/RecommendCard.tsx from recommend.tsx (326→244L)
