+++ hooks/usePhotoPermissions.ts
@@ -1,0 +1,39 @@
+import { useCallback } from "react";
+import { Alert, Linking } from "react-native";
+import * as ImagePicker from "expo-image-picker";
+
+export function usePhotoPermissions() {
+  const requestCameraPermission = useCallback(async (): Promise<boolean> => {
+    const { status } = await ImagePicker.requestCameraPermissionsAsync();
+    if (status !== "granted") {
+      Alert.alert(
+        "Camera Permission Required",
+        "Please enable camera access in your device settings to take progress photos.",
+        [
+          { text: "Cancel", style: "cancel" },
+          { text: "Open Settings", onPress: () => Linking.openSettings() },
+        ]
+      );
+      return false;
+    }
+    return true;
+  }, []);
+
+  const requestGalleryPermission = useCallback(async (): Promise<boolean> => {
+    const { status } = await ImagePicker.requestMediaLibraryPermissionsAsync();
+    if (status !== "granted") {
+      Alert.alert(
+        "Gallery Permission Required",
+        "Please enable photo library access in your device settings to import progress photos.",
+        [
+          { text: "Cancel", style: "cancel" },
+          { text: "Open Settings", onPress: () => Linking.openSettings() },
+        ]
+      );
+      return false;
+    }
+    return true;
+  }, []);
+
+  return { requestCameraPermission, requestGalleryPermission };
+}
